added abstract WhereConditionsBaseDirective class
