+++ src/WhereConditions/WhereConditionsBaseDirective.php
@@ -1,0 +1,178 @@
+<?php
+
+namespace Nuwave\Lighthouse\WhereConditions;
+
+use GraphQL\Error\Error;
+use GraphQL\Language\AST\FieldDefinitionNode;
+use GraphQL\Language\AST\InputValueDefinitionNode;
+use GraphQL\Language\AST\ObjectTypeDefinitionNode;
+use Illuminate\Support\Str;
+use Nuwave\Lighthouse\Schema\AST\Codegen;
+use Nuwave\Lighthouse\Schema\AST\DocumentAST;
+use Nuwave\Lighthouse\Schema\AST\PartialParser;
+use Nuwave\Lighthouse\Schema\Directives\BaseDirective;
+use Nuwave\Lighthouse\Support\Contracts\ArgBuilderDirective;
+use Nuwave\Lighthouse\Support\Contracts\ArgManipulator;
+use Nuwave\Lighthouse\Support\Contracts\DefinedDirective;
+
+abstract class WhereConditionsBaseDirective extends BaseDirective implements ArgBuilderDirective, ArgManipulator, DefinedDirective
+{
+    /**
+     * @var \Nuwave\Lighthouse\WhereConditions\Operator
+     */
+    protected $operator;
+
+    /**
+     * WhereConditions constructor.
+     *
+     * @param  \Nuwave\Lighthouse\WhereConditions\Operator  $operator
+     * @return void
+     */
+    public function __construct(Operator $operator)
+    {
+        $this->operator = $operator;
+    }
+
+    /**
+     * @param  \Illuminate\Database\Query\Builder|\Illuminate\Database\Eloquent\Builder  $builder
+     * @param  array  $whereConditions
+     * @return \Illuminate\Database\Query\Builder|\Illuminate\Database\Eloquent\Builder
+     */
+    public function handleWhereHasConditions($builder, array $whereConditions)
+    {
+        // Make sure to ignore empty conditions.
+        // The "operator" key set by default, so the count is 1 when the condition is empty.
+        if (count($whereConditions) > 1) {
+            $relationName = $this->directiveArgValue('relation')
+                // Use the name of the argument if no explicit relation name is given.
+                ?? $this->nodeName();
+
+            $builder->whereHas(
+                $relationName,
+                function ($builder) use ($whereConditions): void {
+                    // This extra nesting is required for the `OR` condition to work correctly.
+                    $builder->whereNested(
+                        function ($builder) use ($whereConditions): void {
+                            $this->handleWhereConditions($builder, $whereConditions);
+                        }
+                    );
+                }
+            );
+        }
+
+        return $builder;
+    }
+
+    /**
+     * @param  \Illuminate\Database\Query\Builder|\Illuminate\Database\Eloquent\Builder  $builder
+     * @param  array  $whereConditions
+     * @param  string  $boolean
+     * @return \Illuminate\Database\Query\Builder|\Illuminate\Database\Eloquent\Builder
+     */
+    public function handleWhereConditions($builder, array $whereConditions, string $boolean = 'and')
+    {
+        if ($andConnectedConditions = $whereConditions['AND'] ?? null) {
+            $builder->whereNested(
+                function ($builder) use ($andConnectedConditions): void {
+                    foreach ($andConnectedConditions as $condition) {
+                        $this->handleWhereConditions($builder, $condition);
+                    }
+                }
+            );
+        }
+
+        if ($orConnectedConditions = $whereConditions['OR'] ?? null) {
+            $builder->whereNested(
+                function ($builder) use ($orConnectedConditions): void {
+                    foreach ($orConnectedConditions as $condition) {
+                        $this->handleWhereConditions($builder, $condition, 'or');
+                    }
+                },
+                'or'
+            );
+        }
+
+        if ($column = $whereConditions['column'] ?? null) {
+            static::assertValidColumnName($column);
+
+            return $this->operator->applyConditions($builder, $whereConditions, $boolean);
+        }
+
+        return $builder;
+    }
+
+    public static function invalidColumnName(string $column): string
+    {
+        return "Column names may contain only alphanumerics or underscores, and may not begin with a digit, got: $column";
+    }
+
+    /**
+     * Manipulate the AST.
+     *
+     * @param  \Nuwave\Lighthouse\Schema\AST\DocumentAST  $documentAST
+     * @param  \GraphQL\Language\AST\InputValueDefinitionNode  $argDefinition
+     * @param  \GraphQL\Language\AST\FieldDefinitionNode  $parentField
+     * @param  \GraphQL\Language\AST\ObjectTypeDefinitionNode  $parentType
+     * @return void
+     */
+    public function manipulateArgDefinition(
+        DocumentAST &$documentAST,
+        InputValueDefinitionNode &$argDefinition,
+        FieldDefinitionNode &$parentField,
+        ObjectTypeDefinitionNode &$parentType
+    ): void {
+        if ($allowedColumns = $this->directiveArgValue('columns')) {
+            $restrictedWhereConditionsName = $this->restrictedWhereConditionsName($argDefinition, $parentField);
+            $argDefinition->type = PartialParser::namedType($restrictedWhereConditionsName);
+
+            $allowedColumnsEnumName = Codegen::allowedColumnsEnumName($argDefinition, $parentField);
+
+            $documentAST
+                ->setTypeDefinition(
+                    WhereConditionsServiceProvider::createWhereConditionsInputType(
+                        $restrictedWhereConditionsName,
+                        "Dynamic relationship WHERE conditions for the `{$argDefinition->name->value}` argument on the query `{$parentField->name->value}`.",
+                        $allowedColumnsEnumName
+                    )
+                )
+                ->setTypeDefinition(
+                    Codegen::createAllowedColumnsEnum($argDefinition, $parentField, $allowedColumns, $allowedColumnsEnumName)
+                );
+        } else {
+            $argDefinition->type = PartialParser::namedType(WhereConditionsServiceProvider::DEFAULT_WHERE_CONDITIONS);
+        }
+    }
+
+    /**
+     * Create the name for the restricted WhereConditions input.
+     *
+     * @example FieldNameArgNameWhereHasConditions
+     *
+     * @param  \GraphQL\Language\AST\InputValueDefinitionNode  $argDefinition
+     * @param  \GraphQL\Language\AST\FieldDefinitionNode  $parentField
+     * @return string
+     */
+    protected function restrictedWhereConditionsName(InputValueDefinitionNode &$argDefinition, FieldDefinitionNode &$parentField): string
+    {
+        return Str::studly($parentField->name->value)
+            .Str::studly($argDefinition->name->value)
+            .'WhereConditions';
+    }
+
+    /**
+     * Ensure the column name is well formed and prevent SQL injection.
+     *
+     * @param  string  $column
+     * @return void
+     *
+     * @throws \GraphQL\Error\Error
+     */
+    protected static function assertValidColumnName(string $column): void
+    {
+        if (! \Safe\preg_match('/^(?![0-9])[A-Za-z0-9_-]*$/', $column)) {
+            throw new Error(
+                self::invalidColumnName($column)
+            );
+        }
+    }
+}
